+++ docs/changelog.md
@@ -1,5 +1,9 @@
 # Changelog
 
+### 31-March-2022 - 11:55 CEST
+
+- [config] Upgrade Conan client version to 1.45.0.
+
 ### 25-March-2022 - 09:54 CET
 
 - [feature] Updated CMake version to 3.20 only in MacOS machines.